Reorganized and renamed programs
Done by Hanson C.

- Refactored programs to correct folders and names
